Time: 228 ms (31.35%), Space: 46.1 MB (66.47%) - LeetHub

+++ 985-sum-of-even-numbers-after-queries/985-sum-of-even-numbers-after-queries.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     vector<int> sumEvenAfterQueries(vector<int>& nums, vector<vector<int>>& queries) {
         vector<int> ans;
-        long sum =0;
+        int sum =0;
         
          int n = nums.size();
